api: promote Zod↔Convex validation registry to src/lib/validations (#998)
Promotes the pairing table out of convex/validationDrift.test.ts into
src/lib/validations/registry.ts (R-3.1), so both the drift test AND the
Phase 2 dispatcher import ONE table instead of the dispatcher growing a
second hand-maintained copy. Adds OPERATION_VALIDATION_PAIR, a best-effort
"<module>.<fn>" -> pair-name map covering the ~24 create/update operations
that have a curated Zod schema, for the dispatcher's business-field
validation pass. An operation absent from the map isn't unsafe — it falls
back to the in-Convex fieldGuards/moneyGuards backstop, same as every
existing write.
