+++ convex/validationDrift.test.ts
@@ -12,166 +12,16 @@
  * removing a field on one side without the other is drift — this test fails until
  * they're reconciled, which is the guarantee R-8.6.1 asks for. Per-pair `allow`
  * lists document any deliberate, reviewed divergence.
+ *
+ * The pairing table itself lives in src/lib/validations/registry.ts (promoted
+ * there in Phase 2 / #998, R-3.1) so the API dispatcher can run the exact same
+ * Zod↔Convex parity check the UI hooks get — this file only asserts it holds.
  */
 import { describe, expect, it } from "vitest";
-
-import { clientFields } from "./clientWrites";
-import { clientContactFields } from "./clientContactWrites";
-import { assignmentFields } from "./crewAssignmentsWrites";
-import { entryFields } from "./crewTimeEntriesWrites";
-import { checkRecordFields } from "./checkRecords";
-import { categoryFields } from "./categoriesWrites";
-import { checkItemFields } from "./checkItemsWrites";
-import { bulkFields } from "./bulkAssetsWrites";
-import { prefFields } from "./userNotificationPreferences";
-import { modelFields } from "./modelWrites";
-import { locationFields } from "./locationsWrites";
-import { supplierFields } from "./suppliersWrites";
-import { subTestFields } from "./subTestRecords";
-import { projectWriteFields } from "./projects";
-import { updateFields as supplierOrderUpdateFields } from "./supplierOrdersWrites";
-import { itemFields as supplierOrderItemFields } from "./supplierOrderItemsWrites";
-import { quoteSendFields, quoteRecallFields, quoteAcceptFields, quoteDeclineFields } from "./quotesWrites";
-import { invoiceFields } from "./invoicesWrites";
-
-import { clientSchema } from "@/lib/validations/client";
-import { clientContactSchema } from "@/lib/validations/client-contact";
-import { projectSchema } from "@/lib/validations/project";
-import { crewAssignmentSchema, crewTimeEntrySchema } from "@/lib/validations/crew";
-import { checkRecordSchema, checkItemSchema } from "@/lib/validations/check-item";
-import { categorySchema } from "@/lib/validations/category";
-import { bulkAssetSchema, locationSchema } from "@/lib/validations/asset";
-import { notificationPreferenceSchema } from "@/lib/validations/notification-preferences";
-import { modelSchema } from "@/lib/validations/model";
-import { supplierSchema } from "@/lib/validations/supplier";
-import { subTestRecordSchema } from "@/lib/validations/test-tag";
-import { supplierOrderUpdateSchema, supplierOrderItemSchema } from "@/lib/validations/supplier-order";
-import {
-  quoteSendSchema,
-  quoteRecallSchema,
-  quoteAcceptSchema,
-  quoteDeclineSchema,
-} from "@/lib/validations/quote";
-import { invoiceSchema } from "@/lib/validations/invoice";
-
-/** Unwrap a Zod schema (through .refine/.default/.optional wrappers) to its object shape keys. */
-function zodKeys(schema: unknown): string[] {
-  type Wrap = { schema?: unknown; innerType?: unknown; in?: unknown };
-  type Node = { shape?: Record<string, unknown>; _def?: Wrap; def?: Wrap };
-  let cur = schema as Node | null | undefined;
-  for (let i = 0; i < 8 && cur; i++) {
-    if (cur.shape) return Object.keys(cur.shape);
-    const def = cur._def ?? cur.def;
-    cur = (def?.schema ?? def?.innerType ?? def?.in ?? null) as Node | null;
-  }
-  throw new Error("could not resolve object shape from schema");
-}
-
-interface Pair {
-  name: string;
-  zod: unknown;
-  convex: Record<string, unknown>;
-  /** Keys deliberately present on only one side (reviewed). */
-  allowZodOnly?: string[];
-  allowConvexOnly?: string[];
-}
-
-const PAIRS: Pair[] = [
-  { name: "client", zod: clientSchema, convex: clientFields },
-  { name: "clientContact", zod: clientContactSchema, convex: clientContactFields },
-  {
-    name: "crewAssignment",
-    zod: crewAssignmentSchema,
-    convex: assignmentFields,
-    // `generateShifts` is a client-only form control (asks the server to spawn
-    // shifts); it isn't a persisted assignment field.
-    allowZodOnly: ["generateShifts"],
-  },
-  { name: "crewTimeEntry", zod: crewTimeEntrySchema, convex: entryFields },
-  {
-    name: "checkRecord",
-    zod: checkRecordSchema,
-    convex: checkRecordFields,
-    // The client submits { checkItemId, notes, photos, result, value }; the server
-    // record additionally carries identity, the resolved target refs, denormalised
-    // snapshots, and audit stamps — all server-managed, never client input.
-    allowConvexOnly: [
-      "assetId", "bulkAssetId", "checkItemLabelSnapshot", "checkItemTypeSnapshot",
-      "context", "id", "kitId", "lineItemId", "lineItemUnitId", "organizationId",
-      "performedAt", "performedById",
-    ],
-  },
-  { name: "category", zod: categorySchema, convex: categoryFields },
-  { name: "checkItem", zod: checkItemSchema, convex: checkItemFields },
-  { name: "bulkAsset", zod: bulkAssetSchema, convex: bulkFields },
-  { name: "notificationPreference", zod: notificationPreferenceSchema, convex: prefFields },
-  {
-    name: "model",
-    zod: modelSchema,
-    convex: modelFields,
-    // WS6 #945: maintenanceIntervalDays was removed from the interactive
-    // create/edit form's Zod schema (superseded by serviceSchedules) but is
-    // deliberately LEFT wired in modelWrites.ts because CSV bulk import/export
-    // (src/server/csv.ts) still reads/writes this column via the generic
-    // convex/models.ts CRUD — migrating that surface is a separate follow-up.
-    // See the deprecation comment on convex/schema.ts's models.maintenanceIntervalDays.
-    allowConvexOnly: ["maintenanceIntervalDays"],
-  },
-  { name: "location", zod: locationSchema, convex: locationFields },
-  { name: "supplier", zod: supplierSchema, convex: supplierFields },
-  {
-    name: "subTestRecord",
-    zod: subTestRecordSchema,
-    convex: subTestFields,
-    // id + createdAt + the parent FK are server-managed.
-    allowConvexOnly: ["createdAt", "id", "testTagRecordId"],
-  },
-  {
-    name: "project",
-    zod: projectSchema,
-    convex: projectWriteFields,
-    // name + projectNumber are top-level createNative args (not part of the shared
-    // projectWriteFields spread — projectNumber has its own auto-number/clash-guard
-    // handling; name is a required createNative arg, not optional like the rest of
-    // projectWriteFields).
-    allowZodOnly: ["name", "projectNumber"],
-    // Server-managed: the recalc-owned money anchors (never client input, see
-    // PROJECT_MONEY_ANCHORS in projectWrites.ts), isTemplate (set at create, never
-    // patched in place), audit timestamps, and projectManagerId (managed via the
-    // separate projectManagers join table writes, not this field).
-    allowConvexOnly: [
-      "equipmentRevenue", "serviceCostTotal", "labourCostTotal", "subHireCostTotal",
-      "margin", "subtotal", "discountAmount", "taxAmount", "total",
-      // WS1 (#940) — depositPaid/invoicedTotal moved from "hand-typed input" to
-      // recalc-owned PROJECT_MONEY_ANCHORS (derived from the project's Invoice
-      // rows), same treatment as the anchors just above — never a form field.
-      "depositPaid", "invoicedTotal",
-      "isTemplate", "createdAt", "updatedAt", "projectManagerId",
-    ],
-  },
-  // WS7 #946 — supplierOrder HEADER EDIT (status/orderDate/expectedDate/notes only;
-  // supplierId/orderNumber/type/projectId are create-time-only, not part of this pair).
-  { name: "supplierOrderUpdate", zod: supplierOrderUpdateSchema, convex: supplierOrderUpdateFields },
-  // WS7 #946 — supplierOrderItem CRUD (new browser path; previously had none).
-  { name: "supplierOrderItem", zod: supplierOrderItemSchema, convex: supplierOrderItemFields },
-  // WS1 #940 / #986 — the four quote verbs that take client input (new-version
-  // takes none) + invoice create (finance model).
-  { name: "quoteSend", zod: quoteSendSchema, convex: quoteSendFields },
-  { name: "quoteRecall", zod: quoteRecallSchema, convex: quoteRecallFields },
-  { name: "quoteAccept", zod: quoteAcceptSchema, convex: quoteAcceptFields },
-  { name: "quoteDecline", zod: quoteDeclineSchema, convex: quoteDeclineFields },
-  {
-    name: "invoice",
-    zod: invoiceSchema,
-    convex: invoiceFields,
-    // dueDate is a Date client-side, a ms-timestamp number over the wire — the
-    // hook converts it, same pattern as every other date field in this codebase
-    // (see "Date Handling" in FEATUREDOCS/28-patterns.md); field NAME matches.
-  },
-];
+import { VALIDATION_PAIRS, zodKeys } from "@/lib/validations/registry";
 
 describe("validation field-set parity (Zod client ↔ Convex server) — R-8.6.1", () => {
-  it.each(PAIRS)("$name: Convex validator and Zod schema share one field set", (pair) => {
+  it.each(VALIDATION_PAIRS)("$name: Convex validator and Zod schema share one field set", (pair) => {
     const zod = new Set(zodKeys(pair.zod));
     const convex = new Set(Object.keys(pair.convex));
     (pair.allowZodOnly ?? []).forEach((k) => zod.delete(k));
@@ -184,7 +34,7 @@ describe("validation field-set parity (Zod client ↔ Convex server) — R-8.6.1
       { zodOnly, convexOnly },
       `Drift in "${pair.name}": these fields exist on only one side. Reconcile the ` +
         `Zod schema (src/lib/validations) and the Convex *Fields validator, or add a ` +
-        `documented allow entry in validationDrift.test.ts.`,
+        `documented allow entry in src/lib/validations/registry.ts.`,
     ).toEqual({ zodOnly: [], convexOnly: [] });
   });
 });
